Narrow the handler state interface and shift the storage-specific fields to the implementation
